Create H264Transport class. Use it in wowza example
